start updating to 23w31a

+++ azalea-protocol/src/packets/game/clientbound_entity_event_packet.rs
@@ -3,6 +3,8 @@ use azalea_protocol_macros::ClientboundGamePacket;
 
 #[derive(Clone, Debug, McBuf, ClientboundGamePacket)]
 pub struct ClientboundEntityEventPacket {
-    pub entity_id: u32,
-    pub event_id: u8,
+#[var]
+pub entity_id: u32,
+#[var]
+pub event_id: u32,
 }
